course-code/labs/lab-13/solution/tools/insurance_check/requirements.txt
mcp[cli]==1.27.0
fastmcp>=2.0,<3.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.28.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
pytest==8.3.3
pytest-asyncio==0.24.0
